stats: move lib/stat into stat-tags
