+++ connection.js
@@ -25,7 +25,7 @@ var bufrw = require('bufrw');
 var extend = require('xtend');
 var ReadMachine = require('bufrw/stream/read_machine');
 var inherits = require('util').inherits;
-var stat = require('./lib/stat.js');
+var stat = require('./stat-tags.js');
 
 var v2 = require('./v2');
 var errors = require('./errors');
