+++ v2/handler.js
@@ -22,7 +22,7 @@
 
 var EventEmitter = require('../lib/event_emitter');
 var Buffer = require('buffer').Buffer;
-var stat = require('../lib/stat.js');
+var stat = require('../stat-tags.js');
 var util = require('util');
 var assert = require('assert');
